Update tasks to return integers

+++ tasks/tasks.py
@@ -12,7 +12,7 @@ def http_task(self, number):
     for i in range(number):
         time.sleep(.1)
         progress_recorder.set_progress(i+1, number)
-    return random()*1000
+    return int(random()*1000)
 
 
 @shared_task(bind=True)
@@ -32,7 +32,7 @@ def ws_task(self, number):
     for i in range(number):
         time.sleep(.1)
         progress_recorder.set_progress(i+1, number)
-    return random()*1000
+    return int(random()*1000)
 
 
 @shared_task(bind=True)
@@ -43,5 +43,5 @@ def ws_error_task(self, number):
         progress_recorder.set_progress(i+1, number)
         if i == int(number/2):
             raise StopIteration('We broke it!')
-    return random()*1000
+    return int(random()*1000)
